refactor: split 4 more large files by concern
custom_settings.go (1478 → 3 files):
- custom_settings.go: types, CRUD, queries
- custom_settings_secrets.go: secret setting methods
- custom_settings_user.go: user setting methods

functions/storage.go (1191 → 4 files):
- storage.go: function CRUD
- storage_sync.go: sync/import methods
- storage_executions.go: execution logging
- storage_files.go: shared modules and file management

branching/storage.go (1180 → 3 files):
- storage.go: branch CRUD
- storage_queries.go: list, count, query methods
- storage_lifecycle.go: activity, migrations, GitHub, access

oauth_handler.go (1172 → 3 files):
- oauth_handler.go: authorize, providers, state
- oauth_handler_callback.go: callback, user linking
- oauth_handler_providers.go: provider config, user info
