+++ internal/branching/storage.go
@@ -2,7 +2,6 @@ package branching
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"regexp"
@@ -13,7 +12,6 @@ import (
 	"github.com/jackc/pgx/v5"
 	"github.com/jackc/pgx/v5/pgxpool"
 
-	"github.com/nimbleflux/fluxbase/internal/crypto"
 	"github.com/nimbleflux/fluxbase/internal/database"
 )
 
@@ -205,148 +203,6 @@ func (s *Storage) GetBranchByGitHubPR(ctx context.Context, repo string, prNumber
 	return branch, nil
 }
 
-// GetMainBranch retrieves the main branch
-func (s *Storage) GetMainBranch(ctx context.Context) (*Branch, error) {
-	query := `
-		SELECT id, name, slug, database_name, status, type, tenant_id, parent_branch_id,
-			data_clone_mode, github_pr_number, github_pr_url, github_repo,
-			error_message, created_by, created_at, updated_at, expires_at
-		FROM branching.branches
-		WHERE type = 'main' AND status != 'deleted'
-		LIMIT 1`
-
-	branch := &Branch{}
-	err := s.WithTenant(ctx, func(tx pgx.Tx) error {
-		return tx.QueryRow(ctx, query).Scan(
-			&branch.ID,
-			&branch.Name,
-			&branch.Slug,
-			&branch.DatabaseName,
-			&branch.Status,
-			&branch.Type,
-			&branch.TenantID,
-			&branch.ParentBranchID,
-			&branch.DataCloneMode,
-			&branch.GitHubPRNumber,
-			&branch.GitHubPRURL,
-			&branch.GitHubRepo,
-			&branch.ErrorMessage,
-			&branch.CreatedBy,
-			&branch.CreatedAt,
-			&branch.UpdatedAt,
-			&branch.ExpiresAt,
-		)
-	})
-	if errors.Is(err, pgx.ErrNoRows) {
-		return nil, ErrBranchNotFound
-	}
-	if err != nil {
-		return nil, fmt.Errorf("failed to get main branch: %w", err)
-	}
-	return branch, nil
-}
-
-// ListBranches lists branches with optional filtering
-func (s *Storage) ListBranches(ctx context.Context, filter ListBranchesFilter) ([]*Branch, error) {
-	query := `
-		SELECT id, name, slug, database_name, status, type, tenant_id, parent_branch_id,
-			data_clone_mode, github_pr_number, github_pr_url, github_repo,
-			error_message, created_by, created_at, updated_at, expires_at
-		FROM branching.branches
-		WHERE status != 'deleted'`
-
-	args := []any{}
-	argCounter := 1
-
-	if filter.TenantID != nil {
-		query += fmt.Sprintf(" AND tenant_id = $%d", argCounter)
-		args = append(args, *filter.TenantID)
-		argCounter++
-	}
-
-	if filter.Status != nil {
-		query += fmt.Sprintf(" AND status = $%d", argCounter)
-		args = append(args, *filter.Status)
-		argCounter++
-	}
-
-	if filter.Type != nil {
-		query += fmt.Sprintf(" AND type = $%d", argCounter)
-		args = append(args, *filter.Type)
-		argCounter++
-	}
-
-	if filter.CreatedBy != nil {
-		query += fmt.Sprintf(" AND created_by = $%d", argCounter)
-		args = append(args, *filter.CreatedBy)
-		argCounter++
-	}
-
-	if filter.GitHubRepo != nil {
-		query += fmt.Sprintf(" AND github_repo = $%d", argCounter)
-		args = append(args, *filter.GitHubRepo)
-		argCounter++
-	}
-
-	query += " ORDER BY created_at DESC"
-
-	// Use parameterized queries for LIMIT and OFFSET to prevent SQL injection
-	if filter.Limit > 0 {
-		query += fmt.Sprintf(" LIMIT $%d", argCounter)
-		args = append(args, filter.Limit)
-		argCounter++
-	}
-
-	if filter.Offset > 0 {
-		query += fmt.Sprintf(" OFFSET $%d", argCounter)
-		args = append(args, filter.Offset)
-		argCounter++ //nolint:ineffassign // keeping for consistency
-	}
-
-	var branches []*Branch
-	err := s.WithTenant(ctx, func(tx pgx.Tx) error {
-		rows, err := tx.Query(ctx, query, args...)
-		if err != nil {
-			return fmt.Errorf("failed to list branches: %w", err)
-		}
-		defer rows.Close()
-
-		for rows.Next() {
-			branch := &Branch{}
-			err := rows.Scan(
-				&branch.ID,
-				&branch.Name,
-				&branch.Slug,
-				&branch.DatabaseName,
-				&branch.Status,
-				&branch.Type,
-				&branch.TenantID,
-				&branch.ParentBranchID,
-				&branch.DataCloneMode,
-				&branch.GitHubPRNumber,
-				&branch.GitHubPRURL,
-				&branch.GitHubRepo,
-				&branch.ErrorMessage,
-				&branch.CreatedBy,
-				&branch.CreatedAt,
-				&branch.UpdatedAt,
-				&branch.ExpiresAt,
-			)
-			if err != nil {
-				return fmt.Errorf("failed to scan branch: %w", err)
-			}
-			branches = append(branches, branch)
-		}
-
-		return rows.Err()
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	return branches, nil
-}
-
 // UpdateBranchStatus updates the status of a branch
 func (s *Storage) UpdateBranchStatus(ctx context.Context, id uuid.UUID, status BranchStatus, errorMessage *string) error {
 	query := `
@@ -396,613 +252,54 @@ func (s *Storage) DeleteBranch(ctx context.Context, id uuid.UUID, tenantID *uuid
 	})
 }
 
-// CountBranches counts branches matching the filter
-func (s *Storage) CountBranches(ctx context.Context, filter ListBranchesFilter) (int, error) {
-	query := `SELECT COUNT(*) FROM branching.branches WHERE status != 'deleted'`
-
-	args := []any{}
-	argCounter := 1
-
-	if filter.TenantID != nil {
-		query += fmt.Sprintf(" AND tenant_id = $%d", argCounter)
-		args = append(args, *filter.TenantID)
-		argCounter++
-	}
-
-	if filter.Status != nil {
-		query += fmt.Sprintf(" AND status = $%d", argCounter)
-		args = append(args, *filter.Status)
-		argCounter++
-	}
-
-	if filter.Type != nil {
-		query += fmt.Sprintf(" AND type = $%d", argCounter)
-		args = append(args, *filter.Type)
-		argCounter++
-	}
-
-	if filter.CreatedBy != nil {
-		query += fmt.Sprintf(" AND created_by = $%d", argCounter)
-		args = append(args, *filter.CreatedBy)
-	}
-
-	var count int
-	err := s.WithTenant(ctx, func(tx pgx.Tx) error {
-		return tx.QueryRow(ctx, query, args...).Scan(&count)
-	})
-	if err != nil {
-		return 0, fmt.Errorf("failed to count branches: %w", err)
-	}
-
-	return count, nil
-}
-
-// CountBranchesByUser counts branches created by a specific user
-func (s *Storage) CountBranchesByUser(ctx context.Context, userID uuid.UUID) (int, error) {
-	query := `SELECT COUNT(*) FROM branching.branches WHERE created_by = $1 AND status NOT IN ('deleted', 'deleting')`
-
-	var count int
-	err := s.WithTenant(ctx, func(tx pgx.Tx) error {
-		return tx.QueryRow(ctx, query, userID).Scan(&count)
-	})
-	if err != nil {
-		return 0, fmt.Errorf("failed to count user branches: %w", err)
-	}
-
-	return count, nil
-}
-
-// LogActivity records an activity log entry
-func (s *Storage) LogActivity(ctx context.Context, log *ActivityLog) error {
+// SetBranchExpiresAt sets the expiration time for a branch
+func (s *Storage) SetBranchExpiresAt(ctx context.Context, id uuid.UUID, expiresAt *time.Time) error {
 	query := `
-		INSERT INTO branching.activity_log (
-			id, branch_id, tenant_id, action, status, details, error_message, executed_by, duration_ms
-		) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9)
-		RETURNING executed_at`
-
-	if log.ID == uuid.Nil {
-		log.ID = uuid.New()
-	}
-
-	var detailsJSON []byte
-	if log.Details != nil {
-		var err error
-		detailsJSON, err = json.Marshal(log.Details)
-		if err != nil {
-			return fmt.Errorf("failed to marshal details: %w", err)
-		}
-	}
+		UPDATE branching.branches
+		SET expires_at = $1, updated_at = NOW()
+		WHERE id = $2`
 
 	return s.WithTenant(ctx, func(tx pgx.Tx) error {
-		return tx.QueryRow(
-			ctx, query,
-			log.ID,
-			log.BranchID,
-			log.TenantID,
-			log.Action,
-			log.Status,
-			detailsJSON,
-			log.ErrorMessage,
-			log.ExecutedBy,
-			log.DurationMs,
-		).Scan(&log.ExecutedAt)
-	})
-}
-
-// GetActivityLog retrieves activity logs for a branch
-func (s *Storage) GetActivityLog(ctx context.Context, branchID uuid.UUID, limit int) ([]*ActivityLog, error) {
-	if limit <= 0 {
-		limit = 50
-	}
-
-	query := `
-		SELECT id, branch_id, tenant_id, action, status, details, error_message, executed_by, executed_at, duration_ms
-		FROM branching.activity_log
-		WHERE branch_id = $1
-		ORDER BY executed_at DESC
-		LIMIT $2`
-
-	var logs []*ActivityLog
-	err := s.WithTenant(ctx, func(tx pgx.Tx) error {
-		rows, err := tx.Query(ctx, query, branchID, limit)
+		result, err := tx.Exec(ctx, query, expiresAt, id)
 		if err != nil {
-			return fmt.Errorf("failed to get activity log: %w", err)
-		}
-		defer rows.Close()
-
-		for rows.Next() {
-			log := &ActivityLog{}
-			var detailsJSON []byte
-			err := rows.Scan(
-				&log.ID,
-				&log.BranchID,
-				&log.TenantID,
-				&log.Action,
-				&log.Status,
-				&detailsJSON,
-				&log.ErrorMessage,
-				&log.ExecutedBy,
-				&log.ExecutedAt,
-				&log.DurationMs,
-			)
-			if err != nil {
-				return fmt.Errorf("failed to scan activity log: %w", err)
-			}
-			if detailsJSON != nil {
-				if err := json.Unmarshal(detailsJSON, &log.Details); err != nil {
-					return fmt.Errorf("failed to unmarshal details: %w", err)
-				}
-			}
-			logs = append(logs, log)
-		}
-
-		return rows.Err()
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	return logs, nil
-}
-
-// RecordMigration records a migration applied to a branch
-func (s *Storage) RecordMigration(ctx context.Context, branchID uuid.UUID, version int64, name string) error {
-	query := `
-		INSERT INTO branching.migration_history (branch_id, migration_version, migration_name)
-		VALUES ($1, $2, $3)
-		ON CONFLICT (branch_id, migration_version) DO NOTHING`
-
-	return s.WithTenant(ctx, func(tx pgx.Tx) error {
-		_, err := tx.Exec(ctx, query, branchID, version, name)
-		if err != nil {
-			return fmt.Errorf("failed to record migration: %w", err)
-		}
-		return nil
-	})
-}
-
-// GetMigrationHistory retrieves the migration history for a branch
-func (s *Storage) GetMigrationHistory(ctx context.Context, branchID uuid.UUID) ([]*MigrationHistory, error) {
-	query := `
-		SELECT id, branch_id, migration_version, migration_name, applied_at
-		FROM branching.migration_history
-		WHERE branch_id = $1
-		ORDER BY migration_version ASC`
-
-	var history []*MigrationHistory
-	err := s.WithTenant(ctx, func(tx pgx.Tx) error {
-		rows, err := tx.Query(ctx, query, branchID)
-		if err != nil {
-			return fmt.Errorf("failed to get migration history: %w", err)
-		}
-		defer rows.Close()
-
-		for rows.Next() {
-			mh := &MigrationHistory{}
-			err := rows.Scan(
-				&mh.ID,
-				&mh.BranchID,
-				&mh.MigrationVersion,
-				&mh.MigrationName,
-				&mh.AppliedAt,
-			)
-			if err != nil {
-				return fmt.Errorf("failed to scan migration history: %w", err)
-			}
-			history = append(history, mh)
-		}
-
-		return rows.Err()
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	return history, nil
-}
-
-// GetExpiredBranches returns branches that have passed their expiration time
-func (s *Storage) GetExpiredBranches(ctx context.Context) ([]*Branch, error) {
-	query := `
-		SELECT id, name, slug, database_name, status, type, tenant_id, parent_branch_id,
-			data_clone_mode, github_pr_number, github_pr_url, github_repo,
-			error_message, created_by, created_at, updated_at, expires_at
-		FROM branching.branches
-		WHERE expires_at IS NOT NULL
-			AND expires_at < NOW()
-			AND status NOT IN ('deleted', 'deleting')
-			AND type != 'main'`
-
-	var branches []*Branch
-	err := s.WithTenant(ctx, func(tx pgx.Tx) error {
-		rows, err := tx.Query(ctx, query)
-		if err != nil {
-			return fmt.Errorf("failed to get expired branches: %w", err)
-		}
-		defer rows.Close()
-
-		for rows.Next() {
-			branch := &Branch{}
-			err := rows.Scan(
-				&branch.ID,
-				&branch.Name,
-				&branch.Slug,
-				&branch.DatabaseName,
-				&branch.Status,
-				&branch.Type,
-				&branch.TenantID,
-				&branch.ParentBranchID,
-				&branch.DataCloneMode,
-				&branch.GitHubPRNumber,
-				&branch.GitHubPRURL,
-				&branch.GitHubRepo,
-				&branch.ErrorMessage,
-				&branch.CreatedBy,
-				&branch.CreatedAt,
-				&branch.UpdatedAt,
-				&branch.ExpiresAt,
-			)
-			if err != nil {
-				return fmt.Errorf("failed to scan expired branch: %w", err)
-			}
-			branches = append(branches, branch)
-		}
-
-		return rows.Err()
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	return branches, nil
-}
-
-// GitHub Config methods
-
-// GetGitHubConfig retrieves GitHub config for a repository
-func (s *Storage) GetGitHubConfig(ctx context.Context, repository string) (*GitHubConfig, error) {
-	query := `
-		SELECT id, repository, tenant_id, auto_create_on_pr, auto_delete_on_merge,
-			default_data_clone_mode, webhook_secret, created_at, updated_at
-		FROM branching.github_config
-		WHERE repository = $1`
-
-	config := &GitHubConfig{}
-	var encryptedSecret *string
-	err := s.WithTenant(ctx, func(tx pgx.Tx) error {
-		return tx.QueryRow(ctx, query, repository).Scan(
-			&config.ID,
-			&config.Repository,
-			&config.TenantID,
-			&config.AutoCreateOnPR,
-			&config.AutoDeleteOnMerge,
-			&config.DefaultDataCloneMode,
-			&encryptedSecret,
-			&config.CreatedAt,
-			&config.UpdatedAt,
-		)
-	})
-	if errors.Is(err, pgx.ErrNoRows) {
-		return nil, ErrGitHubConfigNotFound
-	}
-	if err != nil {
-		return nil, fmt.Errorf("failed to get GitHub config: %w", err)
-	}
-
-	if encryptedSecret != nil && *encryptedSecret != "" {
-		decrypted, err := crypto.DecryptWithBytesKey(*encryptedSecret, s.encryptionKey)
-		if err != nil {
-			return nil, fmt.Errorf("failed to decrypt webhook secret: %w", err)
-		}
-		config.WebhookSecret = &decrypted
-	}
-
-	return config, nil
-}
-
-// UpsertGitHubConfig creates or updates GitHub config
-func (s *Storage) UpsertGitHubConfig(ctx context.Context, config *GitHubConfig) error {
-	var encryptedSecret *string
-	if config.WebhookSecret != nil && *config.WebhookSecret != "" {
-		encrypted, err := crypto.EncryptWithBytesKey(*config.WebhookSecret, s.encryptionKey)
-		if err != nil {
-			return fmt.Errorf("failed to encrypt webhook secret: %w", err)
-		}
-		encryptedSecret = &encrypted
-	}
-
-	query := `
-		INSERT INTO branching.github_config (
-			id, repository, tenant_id, auto_create_on_pr, auto_delete_on_merge,
-			default_data_clone_mode, webhook_secret
-		) VALUES ($1, $2, $3, $4, $5, $6, $7)
-		ON CONFLICT (repository, tenant_id) DO UPDATE SET
-			auto_create_on_pr = EXCLUDED.auto_create_on_pr,
-			auto_delete_on_merge = EXCLUDED.auto_delete_on_merge,
-			default_data_clone_mode = EXCLUDED.default_data_clone_mode,
-			webhook_secret = EXCLUDED.webhook_secret,
-			updated_at = NOW()
-		RETURNING id, created_at, updated_at`
-
-	if config.ID == uuid.Nil {
-		config.ID = uuid.New()
-	}
-
-	return s.WithTenant(ctx, func(tx pgx.Tx) error {
-		return tx.QueryRow(
-			ctx, query,
-			config.ID,
-			config.Repository,
-			config.TenantID,
-			config.AutoCreateOnPR,
-			config.AutoDeleteOnMerge,
-			config.DefaultDataCloneMode,
-			encryptedSecret,
-		).Scan(&config.ID, &config.CreatedAt, &config.UpdatedAt)
-	})
-}
-
-// DeleteGitHubConfig deletes GitHub config for a repository
-func (s *Storage) DeleteGitHubConfig(ctx context.Context, repository string) error {
-	query := `DELETE FROM branching.github_config WHERE repository = $1`
-
-	return s.WithTenant(ctx, func(tx pgx.Tx) error {
-		result, err := tx.Exec(ctx, query, repository)
-		if err != nil {
-			return fmt.Errorf("failed to delete GitHub config: %w", err)
+			return fmt.Errorf("failed to set branch expiration: %w", err)
 		}
 
 		if result.RowsAffected() == 0 {
-			return ErrGitHubConfigNotFound
+			return ErrBranchNotFound
 		}
 
 		return nil
 	})
 }
 
-// ListGitHubConfigs lists all GitHub configurations
-func (s *Storage) ListGitHubConfigs(ctx context.Context, tenantID *uuid.UUID) ([]*GitHubConfig, error) {
-	query := `
-		SELECT id, repository, tenant_id, auto_create_on_pr, auto_delete_on_merge,
-			default_data_clone_mode, webhook_secret, created_at, updated_at
-		FROM branching.github_config`
+// SetPool sets the connection pool (for testing)
+func (s *Storage) SetPool(pool *pgxpool.Pool) {
+	s.pool = pool
+}
 
-	args := []any{}
-	if tenantID != nil {
-		query += " WHERE tenant_id = $1"
-		args = append(args, *tenantID)
-	}
-	query += " ORDER BY repository"
+// GetPool returns the connection pool
+func (s *Storage) GetPool() *pgxpool.Pool {
+	return s.pool
+}
 
-	var configs []*GitHubConfig
-	err := s.WithTenant(ctx, func(tx pgx.Tx) error {
-		rows, err := tx.Query(ctx, query, args...)
-		if err != nil {
-			return fmt.Errorf("failed to list GitHub configs: %w", err)
-		}
-		defer rows.Close()
-
-		for rows.Next() {
-			config := &GitHubConfig{}
-			var encryptedSecret *string
-			err := rows.Scan(
-				&config.ID,
-				&config.Repository,
-				&config.TenantID,
-				&config.AutoCreateOnPR,
-				&config.AutoDeleteOnMerge,
-				&config.DefaultDataCloneMode,
-				&encryptedSecret,
-				&config.CreatedAt,
-				&config.UpdatedAt,
-			)
-			if err != nil {
-				return fmt.Errorf("failed to scan GitHub config: %w", err)
-			}
-
-			if encryptedSecret != nil && *encryptedSecret != "" {
-				decrypted, err := crypto.DecryptWithBytesKey(*encryptedSecret, s.encryptionKey)
-				if err != nil {
-					return fmt.Errorf("failed to decrypt webhook secret: %w", err)
-				}
-				config.WebhookSecret = &decrypted
-			}
-
-			configs = append(configs, config)
-		}
-
-		return rows.Err()
-	})
+// Transaction executes a function within a database transaction
+func (s *Storage) Transaction(ctx context.Context, fn func(tx pgx.Tx) error) error {
+	tx, err := s.pool.Begin(ctx)
 	if err != nil {
-		return nil, err
+		return fmt.Errorf("failed to begin transaction: %w", err)
 	}
+	defer func() { _ = tx.Rollback(ctx) }()
 
-	return configs, nil
-}
-
-// Branch Access methods
-
-// GrantAccess grants a user access to a branch
-func (s *Storage) GrantAccess(ctx context.Context, access *BranchAccess) error {
-	query := `
-		INSERT INTO branching.branch_access (id, branch_id, tenant_id, user_id, access_level, granted_by)
-		VALUES ($1, $2, $3, $4, $5, $6)
-		ON CONFLICT (branch_id, user_id) DO UPDATE SET
-			access_level = EXCLUDED.access_level,
-			granted_by = EXCLUDED.granted_by,
-			granted_at = NOW()
-		RETURNING id, granted_at`
-
-	if access.ID == uuid.Nil {
-		access.ID = uuid.New()
-	}
-
-	return s.WithTenant(ctx, func(tx pgx.Tx) error {
-		return tx.QueryRow(
-			ctx, query,
-			access.ID,
-			access.BranchID,
-			access.TenantID,
-			access.UserID,
-			access.AccessLevel,
-			access.GrantedBy,
-		).Scan(&access.ID, &access.GrantedAt)
-	})
-}
-
-// RevokeAccess revokes a user's access to a branch
-func (s *Storage) RevokeAccess(ctx context.Context, branchID, userID uuid.UUID) error {
-	query := `DELETE FROM branching.branch_access WHERE branch_id = $1 AND user_id = $2`
-
-	return s.WithTenant(ctx, func(tx pgx.Tx) error {
-		_, err := tx.Exec(ctx, query, branchID, userID)
+	if err := fn(tx); err != nil {
 		return err
-	})
-}
-
-// GetBranchAccessList returns all access grants for a branch
-func (s *Storage) GetBranchAccessList(ctx context.Context, branchID uuid.UUID) ([]*BranchAccess, error) {
-	query := `
-		SELECT id, branch_id, tenant_id, user_id, access_level, granted_at, granted_by
-		FROM branching.branch_access
-		WHERE branch_id = $1
-		ORDER BY granted_at DESC`
-
-	var accessList []*BranchAccess
-	err := s.WithTenant(ctx, func(tx pgx.Tx) error {
-		rows, err := tx.Query(ctx, query, branchID)
-		if err != nil {
-			return fmt.Errorf("failed to list branch access: %w", err)
-		}
-		defer rows.Close()
-
-		for rows.Next() {
-			access := &BranchAccess{}
-			if err := rows.Scan(
-				&access.ID,
-				&access.BranchID,
-				&access.TenantID,
-				&access.UserID,
-				&access.AccessLevel,
-				&access.GrantedAt,
-				&access.GrantedBy,
-			); err != nil {
-				return fmt.Errorf("failed to scan branch access: %w", err)
-			}
-			accessList = append(accessList, access)
-		}
-
-		return rows.Err()
-	})
-	if err != nil {
-		return nil, err
 	}
 
-	return accessList, nil
-}
-
-// GetUserAccess returns the access level for a specific user on a branch
-func (s *Storage) GetUserAccess(ctx context.Context, branchID, userID uuid.UUID) (*BranchAccess, error) {
-	query := `
-		SELECT id, branch_id, tenant_id, user_id, access_level, granted_at, granted_by
-		FROM branching.branch_access
-		WHERE branch_id = $1 AND user_id = $2`
-
-	access := &BranchAccess{}
-	err := s.WithTenant(ctx, func(tx pgx.Tx) error {
-		return tx.QueryRow(ctx, query, branchID, userID).Scan(
-			&access.ID,
-			&access.BranchID,
-			&access.TenantID,
-			&access.UserID,
-			&access.AccessLevel,
-			&access.GrantedAt,
-			&access.GrantedBy,
-		)
-	})
-	if errors.Is(err, pgx.ErrNoRows) {
-		return nil, ErrBranchNotFound
-	}
-	if err != nil {
-		return nil, fmt.Errorf("failed to get user access: %w", err)
+	if err := tx.Commit(ctx); err != nil {
+		return fmt.Errorf("failed to commit transaction: %w", err)
 	}
 
-	return access, nil
-}
-
-// HasAccess checks if a user has at least the specified access level to a branch
-func (s *Storage) HasAccess(ctx context.Context, branchID, userID uuid.UUID, minLevel BranchAccessLevel) (bool, error) {
-	var result bool
-	err := s.WithTenant(ctx, func(tx pgx.Tx) error {
-		// First check if user is the creator (always has admin access)
-		var createdBy *uuid.UUID
-		err := tx.QueryRow(
-			ctx,
-			`SELECT created_by FROM branching.branches WHERE id = $1`,
-			branchID,
-		).Scan(&createdBy)
-		if err != nil {
-			return fmt.Errorf("failed to check branch creator: %w", err)
-		}
-
-		if createdBy != nil && *createdBy == userID {
-			result = true
-			return nil
-		}
-
-		// Then check explicit access grants
-		query := `
-			SELECT access_level FROM branching.branch_access
-			WHERE branch_id = $1 AND user_id = $2`
-
-		var accessLevel BranchAccessLevel
-		err = tx.QueryRow(ctx, query, branchID, userID).Scan(&accessLevel)
-		if errors.Is(err, pgx.ErrNoRows) {
-			result = false
-			return nil
-		}
-		if err != nil {
-			return fmt.Errorf("failed to check access: %w", err)
-		}
-
-		// Check if access level is sufficient
-		result = isAccessSufficient(accessLevel, minLevel)
-		return nil
-	})
-	if err != nil {
-		return false, err
-	}
-
-	return result, nil
-}
-
-// isAccessSufficient checks if the granted level meets the minimum required level
-func isAccessSufficient(granted, required BranchAccessLevel) bool {
-	levels := map[BranchAccessLevel]int{
-		BranchAccessRead:  1,
-		BranchAccessWrite: 2,
-		BranchAccessAdmin: 3,
-	}
-	return levels[granted] >= levels[required]
-}
-
-// UserHasAccess checks if a user has access to a branch (any level)
-func (s *Storage) UserHasAccess(ctx context.Context, slug string, userID uuid.UUID) (bool, error) {
-	// Get the branch first (no tenant filter — access check is cross-tenant for admin users)
-	branch, err := s.GetBranchBySlug(ctx, slug, nil)
-	if err != nil {
-		if errors.Is(err, ErrBranchNotFound) {
-			return false, nil
-		}
-		return false, err
-	}
-
-	// Main branch is accessible to all authenticated users
-	if branch.Type == BranchTypeMain {
-		return true, nil
-	}
-
-	return s.HasAccess(ctx, branch.ID, userID, BranchAccessRead)
+	return nil
 }
 
 // Helper functions
@@ -1091,21 +388,6 @@ func ValidateSlug(slug string) error {
 	return nil
 }
 
-// CountBranchesByTenant counts branches for a specific tenant
-func (s *Storage) CountBranchesByTenant(ctx context.Context, tenantID uuid.UUID) (int, error) {
-	query := `SELECT COUNT(*) FROM branching.branches WHERE tenant_id = $1 AND status NOT IN ('deleted', 'deleting')`
-
-	var count int
-	err := s.WithTenant(ctx, func(tx pgx.Tx) error {
-		return tx.QueryRow(ctx, query, tenantID).Scan(&count)
-	})
-	if err != nil {
-		return 0, fmt.Errorf("failed to count tenant branches: %w", err)
-	}
-
-	return count, nil
-}
-
 // GenerateTenantBranchDatabaseName generates a database name for a tenant-scoped branch
 func GenerateTenantBranchDatabaseName(prefix, tenantSlug, branchSlug string) string {
 	// tenantSlug: "acme-corp" or "default"
@@ -1127,54 +409,4 @@ func GenerateTenantBranchDatabaseName(prefix, tenantSlug, branchSlug string) str
 	}
 
 	return name
-}
-
-// SetPool sets the connection pool (for testing)
-func (s *Storage) SetPool(pool *pgxpool.Pool) {
-	s.pool = pool
-}
-
-// GetPool returns the connection pool
-func (s *Storage) GetPool() *pgxpool.Pool {
-	return s.pool
-}
-
-// Transaction executes a function within a database transaction
-func (s *Storage) Transaction(ctx context.Context, fn func(tx pgx.Tx) error) error {
-	tx, err := s.pool.Begin(ctx)
-	if err != nil {
-		return fmt.Errorf("failed to begin transaction: %w", err)
-	}
-	defer func() { _ = tx.Rollback(ctx) }()
-
-	if err := fn(tx); err != nil {
-		return err
-	}
-
-	if err := tx.Commit(ctx); err != nil {
-		return fmt.Errorf("failed to commit transaction: %w", err)
-	}
-
-	return nil
-}
-
-// SetBranchExpiresAt sets the expiration time for a branch
-func (s *Storage) SetBranchExpiresAt(ctx context.Context, id uuid.UUID, expiresAt *time.Time) error {
-	query := `
-		UPDATE branching.branches
-		SET expires_at = $1, updated_at = NOW()
-		WHERE id = $2`
-
-	return s.WithTenant(ctx, func(tx pgx.Tx) error {
-		result, err := tx.Exec(ctx, query, expiresAt, id)
-		if err != nil {
-			return fmt.Errorf("failed to set branch expiration: %w", err)
-		}
-
-		if result.RowsAffected() == 0 {
-			return ErrBranchNotFound
-		}
-
-		return nil
-	})
 }
